Enable loadConf on configs
In other library ([fs2-kafka is a good example][1]), it's possible to load
arbitrary configuration from a Map.
This is very handy in case you want to update configuration without
recompile. In our case, we will also have a lot of boilerplate and
error-prone code to convert a typesafe config to a pulsar4s XXXConfig.
It would be simplier to just pass config with a `additionalProperties`.

[1]: https://fd4s.github.io/fs2-kafka/api/fs2/kafka/ConsumerSettings.html#withProperties(properties:Map[String,String]):fs2.kafka.ConsumerSettings[F,K,V]