Add Pulsar reconsumeLater API
Pulsar has a retry letter topic feature that allows messages
to be pushed to a dedicated topic while waiting to be resent
to the consumer.

This feature is disabled by default and must be enable
when building the consumer with `enableRetry(true)`.
When enabled, if the consumer calls the `reconsumeLater`
method, the message will be pushed to a `-RETRY` topic
and redelivered to the consumer after the period defined
when calling `reconsumeLater`.

The maximum number of time the message can be redelivered
is defined by the `maxRedeliverCount` parameter of the
DeadLetterPolicy of the consumer.

When this limit is reached, the message is sent to the
dead letter topic.

This PR aims to add the reconsumeLater API to the Scala implementation.