Add Pulsar reconsumeLater API
Pulsar has a retry letter topic feature that allows messages
to be pushed to a dedicated topic while waiting to be resent
to the consumer.

This feature is disabled by default and must be enable
when building the consumer with `enableRetry(true)`.
When enabled, if the consumer calls the `reconsumeLater`
method, the message will be pushed to a `-RETRY` topic
and redelivered to the consumer after the period defined
when calling `reconsumeLater`.

The maximum number of time the message can be redelivered
is defined by the `maxRedeliverCount` parameter of the
DeadLetterPolicy of the consumer.

When this limit is reached, the message is sent to the
dead letter topic.

In order to use the reconsumerLater API, we face a main issue:
While others pulsar methods need the MessageId to identify the
message, the reconsumeLater API needs the whole message.
Rebuilding the message based on the ConsumerMessage content
is complicated since we need to create a TopicMessageImpl
for consumers consuming multiple topics instead of a MessageImpl
and the TopicMessageImpl constructor has a package visiblity.
One solution is to expose a method that returns the base pulsar
message in one of the implementation of the ConsumerMessage.
This way, we reduce a bit the coupling between the Pulsar java API
and Pulsar4s as we don't expose the method directly in the
ConsumerMessage trait.

This PR aims to add the reconsumeLater API to the Scala implementation.